[GOB-1215] Resolve filename on every export
Filenames were being resolved when an export was first called and the results saved in the config. When a container kept running the filename stayed as it was first resolved, resulting in an incorrect date in the filename. Now the filename is being resolved everytime and the original function isn’t being overwritten by the result of the resolve.