Fix bug with UniqueFilter
Running the same export twice resulted in an empty export when a UniqueFilter was being used. The set used for checking uniqueness wasn’t cleared after an export is completed therefor the next run no records passed the UniqueFilter. A reset method was added which is run after all records are read from the API