Fix dump for Amsterdam Schema relations
Change of relation names going from public to legacy schema caused the renamed relations
to not be dumped. The new relation name could not be found in API.

Changed the dump logic in such a way that it does not fetch the collection from API,
but just calls API with the new name. API handles translating the new names back to the
legacy names.
Fetching the collection from API is still needed when dump is called with
include_relations=True. Because we would not expect object tables to be renamed from
legacy to public, this should never be a problem. However, to make sure this edge case
will never go unnoticed anymore, a new error log is added as well in case the collection
could not be found when we do need it.